Remove this as it was causing a crash on startup.
It was trying to reference `net.minecraft.client.gui.screen.Screen` which it could not find even though it did exist. Unsure the reason, but assuming it is something to do with the fact that it only output a single jar rather than the 4 it would normally do, so this single one may have been the incorrect type/format.